+++ examples/http_server_longpoll.cpp
@@ -46,7 +46,7 @@ int main(int argc, char *argv[]) {
 	std::cout << "This server responds to requests approximately after 5 seconds" << std::endl;
 	crab::RunLoop runloop;
 
-	crab::SignalStop sig([&]() {
+	crab::Signal sig([&]() {
 		std::cout << "Good bye" << std::endl;
 		crab::RunLoop::current()->cancel();
 	});