+++ include/crab/network_libev.hxx
@@ -60,11 +60,11 @@ CRAB_INLINE void Idle::set_active(bool a) {
 
 CRAB_INLINE bool Idle::is_active() { return impl.is_active(); }
 
-CRAB_INLINE SignalStop::SignalStop(Handler &&cb) : a_handler(std::move(cb)) {}
+CRAB_INLINE Signal::Signal(Handler &&cb, const std::vector<int> &) : a_handler(std::move(cb)) {}
 
-CRAB_INLINE SignalStop::~SignalStop() {}
+CRAB_INLINE Signal::~Signal() {}
 
-CRAB_INLINE bool SignalStop::running_under_debugger() { return false; }
+CRAB_INLINE bool Signal::running_under_debugger() { return false; }
 
 CRAB_INLINE RunLoop::RunLoop() : impl(new ev::dynamic_loop{}) {
 	if (CurrentLoop::instance)