+++ ReleaseNotes.md
@@ -3,6 +3,11 @@
 ### 0.9.2
 
 - example of fair resource-bound API server. TODO - improve example
+- `SignalStop` renamed to `Signal` and now accepts list of signals. Empty list means to detect generic `Ctrl-C` and sorta works on `Linux`, `Mac` and `Windows`.
+- `crab::Thread` to wrap creation of `std::thread` with RunLoop and perform raceless cancellation
+
+*Bugs fixed*
+- `sigprocmask` was changed `pthread_sigmask`, so that `Signal` work in multithreaded apps
 
 ### 0.9.1